+++ CHANGELOG.md
@@ -10,6 +10,7 @@ of [keepachangelog.com](http://keepachangelog.com/).
 
 - Completed [README](./README.md)
 - Completed this initial Changelog
+- renaming functions to match conventions for side-effects
 
 ### Removed